refactor: world path 이동.
