+++ EngineSIU/EngineSIU/Engine/Source/Editor/PropertyEditor/PropertyEditorPanel.cpp
@@ -1,6 +1,6 @@
 #include "PropertyEditorPanel.h"
 
-#include "World.h"
+#include "World/World.h"
 #include "Actors/Player.h"
 #include "Components/LightComponent.h"
 #include "Components/StaticMeshComponent.h"
