+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/SkySphereComponent.cpp
@@ -1,5 +1,4 @@
 #include "SkySphereComponent.h"
-#include "World.h"
 
 
 USkySphereComponent::USkySphereComponent()
