+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/SphereComp.cpp
@@ -1,6 +1,5 @@
 #include "SphereComp.h"
 #include "Engine/Source/Runtime/Core/Math/JungleMath.h"
-#include "Engine/Source/Runtime/Engine/World.h"
 #include "UnrealEd/EditorViewportClient.h"
