+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/UBillboardComponent.cpp
@@ -2,7 +2,7 @@
 #include <DirectXMath.h>
 #include "Define.h"
 #include "QuadTexture.h"
-#include "World.h"
+#include "World/World.h"
 #include "Actors/Player.h"
 #include "LevelEditor/SLevelEditor.h"
 #include "Math/MathUtility.h"
