+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/UText.cpp
@@ -1,6 +1,6 @@
 #include "UText.h"
 
-#include "World.h"
+#include "World/World.h"
 #include "Engine/Source/Editor/PropertyEditor/ShowFlags.h"
 #include "UnrealEd/EditorViewportClient.h"
 #include "LevelEditor/SLevelEditor.h"
